// src/models/player.ts
export interface Player {
    Results: ResultElement[];
    Links:   Links;
}

export interface Links {
    Self: Self;
}

export interface Self {
    AuthorityId:                              string;
    Path:                                     string;
    QueryString:                              string;
    RetryPolicyId:                            string;
    TopicName:                                string;
    AcknowledgementTypeId:                    number;
    AuthenticationLifetimeExtensionSupported: boolean;
}

export interface ResultElement {
    Id:         string;
    ResultCode: number;
    Result:     ResultResult;
}

export interface ResultResult {
    ArenaStats:  Arena;
    PlayerId:    PlayerID;
    SpartanRank: number;
    Xp:          number;
}